sf8-behavior-prediction/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
pandas
